PR feedback: updated renameComponent helper method, implemented in expandable-rename-expandablesection & chipgroup-remove-chipbutton
