+++ packages/eslint-plugin-pf-codemods/lib/rules/chipgroup-remove-chipbutton.js
@@ -3,7 +3,6 @@ const { renameComponent } = require('../helpers');
 // https://github.com/patternfly/patternfly-react/pull/4246
 module.exports = {
   create: renameComponent(
-    { 'ChipButton': 'Button' },
-    node => `${node.name} has been replaced with Button`
+    { 'ChipButton': 'Button' }
   )
 };
